Rename CreateCredential page and move View Credentials to Credentials page
