+++ src/routes/routes.ts
@@ -6,10 +6,10 @@ import DecentralizedIDs from "../pages/DecentralizedIDs";
 import IssueCredential from "../pages/IssueCredential";
 import PresentationDefinitions from "../pages/PresentationDefinitions";
 import ResolveDID from "../experimental/ResolveDID";
-import CreateCredentialTemplate from "../pages/CreateCredentialTemplate";
 import Modal from "../components/Modal";
 import CreateDID from "../components/CreateDID";
 import SetDID from "../experimental/SetDID";
+import Credentials from "../pages/Credentials";
 
 
 export const routesWithCustomData = [
@@ -26,7 +26,7 @@ export const routesWithCustomData = [
     },
     {
         path: 'credentials',
-        component: CreateCredentialTemplate,
+        component: Credentials,
         custom: {
             title: 'Credentials'
         },
